+++ example/lib/main.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/material.dart';
-import 'dart:async';
 
-import 'package:flutter/services.dart';
-import 'package:flutter_stripe/flutter_stripe.dart';
+import 'package:flutter_stripe_sdk/stripe.dart';
+import 'package:flutter_stripe_sdk/customer_session.dart';
 
 void main() => runApp(MyApp());
 
@@ -12,9 +11,32 @@ class MyApp extends StatefulWidget {
 }
 
 class _MyAppState extends State<MyApp> {
+  Stripe _stripe;
+
   @override
   void initState() {
     super.initState();
+
+    _stripe = Stripe('your_stripe_publishable_key');
+
+    // Begin customer session
+    CustomerSession.initCustomerSessionUsingFunction((apiVersion, keyUpdateListener) {
+      // TODO Somehow get ephemeral key for customer form your backend
+
+      // On success
+      keyUpdateListener.onKeyUpdate('your ephemeral key json object');
+
+      // On failure:
+      keyUpdateListener.onKeyUpdateFailure(0, 'Cannot get ephemeral key');
+    });
+  }
+
+  @override
+  void dispose() {
+    // End customer session
+    CustomerSession.endCustomerSession();
+
+    super.dispose();
   }
 
   @override