+++ lib/ephemeral_key_provider.dart
@@ -1,4 +1,4 @@
-import 'package:flutter_stripe/ephemeral_key_update_listener.dart';
+import 'package:flutter_stripe_sdk/ephemeral_key_update_listener.dart';
 
 abstract class EphemeralKeyProvider {
   void createEphemeralKey(String apiVersion, EphemeralKeyUpdateListener keyUpdateListener);